Split iOS cloud sync contracts
